+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Fixed
 - Playback stopping when clicking the title "0 results" of an empty advanced search result
 - Layout being totally broken on right-to-left UI languages like Hebrew or Arabic
+- Media control seek forward/backward buttons not working on Edge
 
 ## 2.4.1 - 2025-10-12